Move SystemTrayMenu to App.Ux
