+++ src/cswm.App/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using cswm.App.Services;
+using cswm.App.Ux;
 using cswm.Arrangement;
 using cswm.Events;
 using cswm.Options;
