+++ src/cswm.App/Services/SystemTrayService.cs
@@ -1,3 +1,4 @@
+using cswm.App.Ux;
 using cswm.Events;
 using cswm.Tracking.Events;
 using Microsoft.Extensions.Logging;
